requirements-dev.txt
aiohttp>=3.7.4
PyYAML>=5.4.1
flake8>=3.6.0
pytest>=6.2.5
pytest-asyncio>=0.17.2
coverage>=5.2.1
pytest-cov>=3.0.0
coveralls>=3.3.0
